requirements.txt
google-adk
litellm
python-dotenv
uvicorn[standard]
fastapi
websockets
sqlalchemy
psycopg2-binary
asyncpg
supabase
astrapy
google-generativeai
